+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `instagram` modifier in `buttonIcon` component,
+- Feature: Add `instagram` modifier in `ButtonIcon` component,
   `InstagramButtonIcon` and `InstagramIcon` components.
 - Feature: Add `k-buttonPadding` helper to manage button paddings.
 - Feature: Update all `Button` components with new design styleguide V2.